SaveManager modified for modularity, working on Data Handling module
Removed code that validated savedata in the SaveManager module, will instead be working on a different module for handling save data structures. This module will be a lot easier and more lenient to modify.

Some work done on the module in question, SaveManager/DataHandler.luau. Mainly just the table of all data since the current template is MASSIVE and took me ages to port over.